Fix bug where cleanup code was cleaned up

+++ python/cudaq/kernel/kernel_decorator.py
@@ -549,7 +549,9 @@ class PyKernelDecorator(object):
         """
         Delete the `ExecutionEngine` cache given by a cache key.
         """
-        cudaq_runtime.delete_cache_execution_engine(key)
+        # Make sure this hasn't already been cleaned up as we're winding down
+        if cudaq_runtime.delete_cache_execution_engine is not None:
+            cudaq_runtime.delete_cache_execution_engine(key)
 
     def resolve_decorator_at_callsite(self, callingMod):
         # Resolve all lifted arguments for `self`.
